requirements.txt
# Core Bot Framework
python-telegram-bot==20.7
python-dotenv==1.0.0

# AI and NLP
openai==1.12.0
anthropic==0.18.1

# Astrology calculations
pyswisseph==2.10.3.2
pytz==2024.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Async support
aiohttp==3.9.3
asyncio==3.4.3

# Data processing
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
loguru==0.7.2

# Web scraping (for horoscopes)
beautifulsoup4==4.12.3
lxml==5.1.0

